Update pam_linotp.py
If server-messages were sent back to the client, the 'Otp:' prompt would be overwritten, resulting in no actual request for user action.
This results in:
bash$ sudo -i
[sudo] password for a_unixuser:
e-mail sent successfully

The system would then hang at the end of 'successfully' waiting for unprompted input. If the user does not know to type their OTP, or presses 'enter' to see if the system is hung, auth fails. 

This change re-appends said prompt to the end of the challenge message, before it is sent back to PAM.

[sudo] password for a_unixuser:
e-mail sent successfully - OTP: